Undistortion and rectification of images

+++ orb_slam3_rewrite/include/external/slam_sensor/slam_data.h
@@ -10,13 +10,13 @@ class SlamData
 private:
   cv::Mat image_left;
   cv::Mat image_right;
-  std::vector<std::vector<double>> imu_readings;
+  std::vector<std::vector<float>> imu_readings;
 
 public:
-  SlamData() { imu_readings = std::vector<std::vector<double>>(); }
+  SlamData() { imu_readings = std::vector<std::vector<float>>(); }
   void setImageLeft(cv::Mat image) { image_left = image; }
   void setImageRight(cv::Mat image) { image_right = image; }
-  void addImuReading(std::vector<double> imuReading) { imu_readings.push_back(imuReading); }
+  void addImuReading(std::vector<float> imuReading) { imu_readings.push_back(imuReading); }
 
   cv::Mat getImageLeft() { return image_left; }
   cv::Mat getImageRight() { return image_right; }
